refactor: Move Models and UI Components to Discovery Module
We've streamlined our project by relocating models and UI
components previously stored in the Core module to the Discovery
module. These elements were previously shared across multiple
modules—course, dashboard, and discovery.
