+++ app/src/main/java/org/openedx/app/MainFragment.kt
@@ -16,7 +16,7 @@ import org.openedx.app.databinding.FragmentMainBinding
 import org.openedx.core.config.Config
 import org.openedx.core.presentation.global.app_upgrade.UpgradeRequiredFragment
 import org.openedx.core.presentation.global.viewBinding
-import org.openedx.dashboard.presentation.dashboard.DashboardFragment
+import org.openedx.dashboard.presentation.DashboardFragment
 import org.openedx.discovery.presentation.DiscoveryNavigator
 import org.openedx.discovery.presentation.DiscoveryRouter
 import org.openedx.discovery.presentation.program.ProgramFragment
