--- core/src/main/java/org/openedx/core/domain/model/CourseList.kt
@@ -1,6 +1,0 @@
-package org.openedx.core.domain.model
-
-data class CourseList(
-    val pagination: Pagination,
-    val results: List<Course>,
-)
