--- dashboard/src/main/java/org/openedx/dashboard/notifier/DashboardEvent.kt
@@ -1,6 +1,0 @@
-package org.openedx.dashboard.notifier
-
-sealed class DashboardEvent {
-    object NavigationToDiscovery : DashboardEvent()
-    object UpdateEnrolledCourses : DashboardEvent()
-}
