+++ dashboard/src/main/java/org/openedx/dashboard/presentation/DashboardAnalytics.kt
@@ -1,4 +1,4 @@
-package org.openedx.dashboard.presentation.dashboard
+package org.openedx.dashboard.presentation
 
 interface DashboardAnalytics {
     fun dashboardCourseClickedEvent(courseId: String, courseName: String)
