+++ dashboard/src/main/java/org/openedx/dashboard/presentation/DashboardUIState.kt
@@ -1,4 +1,4 @@
-package org.openedx.dashboard.presentation.dashboard
+package org.openedx.dashboard.presentation
 
 import org.openedx.core.domain.model.EnrolledCourse
