+++ discovery/src/main/java/org/openedx/discovery/data/model/CourseList.kt
@@ -1,6 +1,7 @@
-package org.openedx.core.data.model
+package org.openedx.discovery.data.model
 
 import com.google.gson.annotations.SerializedName
+import org.openedx.core.data.model.Pagination
 
 data class CourseList(
     @SerializedName("pagination")
