+++ discovery/src/main/java/org/openedx/discovery/domain/interactor/DiscoveryInteractor.kt
@@ -1,8 +1,8 @@
 package org.openedx.discovery.domain.interactor
 
-import org.openedx.core.domain.model.Course
-import org.openedx.core.domain.model.CourseList
 import org.openedx.discovery.data.repository.DiscoveryRepository
+import org.openedx.discovery.domain.model.Course
+import org.openedx.discovery.domain.model.CourseList
 
 class DiscoveryInteractor(private val repository: DiscoveryRepository) {
