+++ discovery/src/main/java/org/openedx/discovery/domain/model/CourseList.kt
@@ -1,0 +1,8 @@
+package org.openedx.discovery.domain.model
+
+import org.openedx.core.domain.model.Pagination
+
+data class CourseList(
+    val pagination: Pagination,
+    val results: List<Course>,
+)
