+++ discovery/src/main/java/org/openedx/discovery/presentation/DiscoveryUIState.kt
@@ -1,8 +1,8 @@
 package org.openedx.discovery.presentation
 
-import org.openedx.core.domain.model.Course
+import org.openedx.discovery.domain.model.Course
 
 sealed class DiscoveryUIState {
     data class Courses(val courses: List<Course>) : DiscoveryUIState()
-    object Loading : DiscoveryUIState()
+    data object Loading : DiscoveryUIState()
 }
