+++ discovery/src/main/java/org/openedx/discovery/presentation/WebViewDiscoveryFragment.kt
@@ -50,9 +50,6 @@ import androidx.lifecycle.LifecycleEventObserver
 import androidx.lifecycle.LifecycleOwner
 import org.koin.androidx.viewmodel.ext.android.viewModel
 import org.koin.core.parameter.parametersOf
-import org.openedx.core.presentation.CoreAnalyticsScreen
-import org.openedx.core.presentation.catalog.CatalogWebViewScreen
-import org.openedx.core.presentation.catalog.WebViewLink
 import org.openedx.core.presentation.dialog.alert.ActionDialogFragment
 import org.openedx.core.ui.AuthButtonsPanel
 import org.openedx.core.ui.ConnectionErrorView
@@ -66,6 +63,8 @@ import org.openedx.core.ui.theme.OpenEdXTheme
 import org.openedx.core.ui.theme.appColors
 import org.openedx.core.ui.windowSizeValue
 import org.openedx.discovery.R
+import org.openedx.discovery.presentation.catalog.CatalogWebViewScreen
+import org.openedx.discovery.presentation.catalog.WebViewLink
 import org.openedx.core.R as CoreR
 
 class WebViewDiscoveryFragment : Fragment() {
@@ -126,7 +125,7 @@ class WebViewDiscoveryFragment : Fragment() {
                                         getString(CoreR.string.platform_name)
                                     ),
                                     url = param,
-                                    source = CoreAnalyticsScreen.DISCOVERY.screenName
+                                    source = DiscoveryAnalyticsScreen.DISCOVERY.screenName
                                 ).show(
                                     requireActivity().supportFragmentManager,
                                     ActionDialogFragment::class.simpleName
