+++ discovery/src/main/java/org/openedx/discovery/presentation/detail/CourseDetailsUIState.kt
@@ -1,10 +1,10 @@
 package org.openedx.discovery.presentation.detail
 
-import org.openedx.core.domain.model.Course
+import org.openedx.discovery.domain.model.Course
 
 sealed class CourseDetailsUIState {
     data class CourseData(val course: Course, val isUserLoggedIn: Boolean = false) :
         CourseDetailsUIState()
 
-    object Loading : CourseDetailsUIState()
+    data object Loading : CourseDetailsUIState()
 }
