+++ discovery/src/main/java/org/openedx/discovery/presentation/program/ProgramFragment.kt
@@ -43,9 +43,6 @@ import androidx.core.os.bundleOf
 import androidx.fragment.app.Fragment
 import org.koin.androidx.viewmodel.ext.android.viewModel
 import org.openedx.core.extension.toastMessage
-import org.openedx.core.presentation.CoreAnalyticsScreen
-import org.openedx.core.presentation.catalog.CatalogWebViewScreen
-import org.openedx.core.presentation.catalog.WebViewLink
 import org.openedx.core.presentation.dialog.alert.ActionDialogFragment
 import org.openedx.core.presentation.dialog.alert.InfoDialogFragment
 import org.openedx.core.ui.ConnectionErrorView
@@ -60,8 +57,11 @@ import org.openedx.core.ui.theme.OpenEdXTheme
 import org.openedx.core.ui.theme.appColors
 import org.openedx.core.ui.windowSizeValue
 import org.openedx.discovery.R
+import org.openedx.discovery.presentation.DiscoveryAnalyticsScreen
+import org.openedx.discovery.presentation.catalog.CatalogWebViewScreen
+import org.openedx.discovery.presentation.catalog.WebViewLink
 import org.openedx.core.R as coreR
-import org.openedx.core.presentation.catalog.WebViewLink.Authority as linkAuthority
+import org.openedx.discovery.presentation.catalog.WebViewLink.Authority as linkAuthority
 
 class ProgramFragment(private val myPrograms: Boolean = false) : Fragment() {
 
@@ -171,14 +171,12 @@ class ProgramFragment(private val myPrograms: Boolean = false) : Fragment() {
                                         getString(coreR.string.platform_name)
                                     ),
                                     url = param,
-                                    source = CoreAnalyticsScreen.PROGRAM.screenName
+                                    source = DiscoveryAnalyticsScreen.PROGRAM.screenName
                                 ).show(
                                     requireActivity().supportFragmentManager,
                                     ActionDialogFragment::class.simpleName
                                 )
                             }
-
-                            else -> {}
                         }
                     },
                     refreshSessionCookie = {
