+++ discovery/src/main/java/org/openedx/discovery/presentation/search/CourseSearchFragment.kt
@@ -64,11 +64,9 @@ import androidx.fragment.app.Fragment
 import org.koin.android.ext.android.inject
 import org.koin.androidx.viewmodel.ext.android.viewModel
 import org.openedx.core.UIMessage
-import org.openedx.core.domain.model.Course
 import org.openedx.core.domain.model.Media
 import org.openedx.core.ui.AuthButtonsPanel
 import org.openedx.core.ui.BackBtn
-import org.openedx.core.ui.DiscoveryCourseItem
 import org.openedx.core.ui.HandleUIMessage
 import org.openedx.core.ui.SearchBar
 import org.openedx.core.ui.WindowSize
@@ -80,7 +78,9 @@ import org.openedx.core.ui.theme.OpenEdXTheme
 import org.openedx.core.ui.theme.appColors
 import org.openedx.core.ui.theme.appTypography
 import org.openedx.core.ui.windowSizeValue
+import org.openedx.discovery.domain.model.Course
 import org.openedx.discovery.presentation.DiscoveryRouter
+import org.openedx.discovery.presentation.ui.DiscoveryCourseItem
 import org.openedx.discovery.R as discoveryR
 
 class CourseSearchFragment : Fragment() {
