+++ discovery/src/main/java/org/openedx/discovery/presentation/search/CourseSearchUIState.kt
@@ -1,8 +1,8 @@
 package org.openedx.discovery.presentation.search
 
-import org.openedx.core.domain.model.Course
+import org.openedx.discovery.domain.model.Course
 
 sealed class CourseSearchUIState {
     data class Courses(val courses: List<Course>, val numCourses: Int) : CourseSearchUIState()
-    object Loading : CourseSearchUIState()
+    data object Loading : CourseSearchUIState()
 }
