+++ discovery/src/test/java/org/openedx/discovery/presentation/search/CourseSearchViewModelTest.kt
@@ -23,12 +23,12 @@ import org.openedx.core.R
 import org.openedx.core.UIMessage
 import org.openedx.core.config.Config
 import org.openedx.core.data.storage.CorePreferences
-import org.openedx.core.domain.model.Course
-import org.openedx.core.domain.model.CourseList
 import org.openedx.core.domain.model.Media
 import org.openedx.core.domain.model.Pagination
 import org.openedx.core.system.ResourceManager
 import org.openedx.discovery.domain.interactor.DiscoveryInteractor
+import org.openedx.discovery.domain.model.Course
+import org.openedx.discovery.domain.model.CourseList
 import org.openedx.discovery.presentation.DiscoveryAnalytics
 import java.net.UnknownHostException
